ArtworkDataset filtering (artist name, artwork title), uri-info (.py and .sh and .nem_action)
- dataset_filternig.py
Filter the dataset throwing away the artwork without a title and an artist name.
This filtering is done because we are making the query to google and flickr using the title and the artist name.

- uri-info.py
Print on terminal the information about the artwork associated with the uri in input.
Can be used with CLI.
Can be used with uri-info.sh.
Can be integrated with Nemo file explorer (fork of Nautilus, default file explorer in cinnamon linux) with the uri_info_nemo_plugin (right click on folder of the artwork in the dataset and select uri-info entry)